agents/animation-worker/pyproject.toml
[project]
name = "animation-worker"
version = "0.1.0"
description = "Animation worker agent using Claude Agent SDK with Gemini vision analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "aiofiles>=23.0",
    "pygithub>=2.0",
    "rich>=13.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0",
    "claude-agent-sdk>=0.1.0",
    "google-genai>=1.0.0",
]

[project.scripts]
animation-worker = "animation_worker.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/animation_worker"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
